Small refactoring for `FutureUnodered` iterators
