+++ futures-util/src/stream/futures_unordered/iter.rs
@@ -12,7 +12,9 @@ pub struct IterPinMut<'a, F: 'a> {
     pub(super) _marker: PhantomData<&'a mut FuturesUnordered<F>>
 }
 
-impl<'a, F: Unpin> Unpin for IterPinMut<'a, F> {}
+#[derive(Debug)]
+/// Mutable iterator over all futures in the unordered set.
+pub struct IterMut<'a, F: 'a + Unpin> (pub(super) IterPinMut<'a, F>);
 
 impl<'a, F> Iterator for IterPinMut<'a, F> {
     type Item = PinMut<'a, F>;
@@ -36,3 +38,17 @@ impl<'a, F> Iterator for IterPinMut<'a, F> {
 }
 
 impl<'a, F> ExactSizeIterator for IterPinMut<'a, F> {}
+
+impl<'a, F: Unpin> Iterator for IterMut<'a, F> {
+    type Item = &'a mut F;
+
+    fn next(&mut self) -> Option<&'a mut F> {
+        self.0.next().map(|f| unsafe { PinMut::get_mut(f) })
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        self.0.size_hint()
+    }
+}
+
+impl<'a, F: Unpin> ExactSizeIterator for IterMut<'a, F> {}
